requirements.txt
spotipy==2.24.0
pandas==2.2.2
numpy==1.26.4
scikit-learn==1.5.0
umap-learn==0.5.6
plotly==5.22.0
streamlit==1.35.0
python-dotenv==1.0.1
joblib==1.4.2
requests==2.32.3
jupyter==1.0.0
ipykernel==6.29.4
matplotlib==3.9.0
seaborn==0.13.2
